add: Log out button
